I have added the LoginActivity adn SignUpActivity java files along with their xml files. They have (as far as I can currently tell) full functionality.
To be done: Must connect MainActivity to be able to pick a button and lead to either option. Also, we must decide which values will remain. I put the common values (used username and no email for now) but we can add or remove values.

Pressing commit and push because I'm not sure if things updated